refactor: registartion page split into smaller component
